Initial commit: Set up React project structure and components
This commit includes the initial setup of a React project with the following components:

App component: The main component that renders the application.
CardList component: Displays a list of cards containing user information.
FilterInput component: Allows filtering the list of cards by name.
AddUser component: Provides a form to add a new user to the list.
Card component: Represents a single user card with their information.